% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{A_phi}
\alias{A_phi}
\title{A_phi}
\usage{
A_phi(T_kelvin, pressure_atm = 1)
}
\arguments{
\item{T_kelvin}{ambient temperature in Kelvin}

\item{pressure_atm}{ambient pressure in atmospheres, default 1}
}
\description{
Get the predicted A_phi parameter for the Pitzer equation
at a given temperature and pressure,
using equations from PHREEQC
}
