% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{buck_eqn}
\alias{buck_eqn}
\title{buck_eqn}
\usage{
buck_eqn(T_kelvin)
}
\arguments{
\item{T_kelvin}{Ambient temperature in Kelvin}
}
\value{
corresponding saturation vapor pressure
}
\description{
returns a value for saturation vapor
pressure in Pa using Buck's empiricial
equation
}
