% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{cohen_molality_of_RH}
\alias{cohen_molality_of_RH}
\title{cohen_molality_of_RH}
\usage{
cohen_molality_of_RH(RH, tolerance = 0.01)
}
\arguments{
\item{RH}{ambient relative humidity}
}
\value{
corresponding molality of NaCl
}
\description{
Get the molality of
NaCl in pure water at env conditions
corresponding to the given RH by numerically
inverting the empirical equation per Cohen et al
https://doi.org/10.1021/j100301a029
}
