% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{cohen_RH_of_molality}
\alias{cohen_RH_of_molality}
\title{cohen_RH_of_molality}
\usage{
cohen_RH_of_molality(molality)
}
\arguments{
\item{molality}{molality of NaCl}
}
\value{
corresponding relative humidity
}
\description{
Get the relative humidity
corresponding to a given molality of
NaCl in pure water at env conditions
from an empirical equation per Cohen et al
https://doi.org/10.1021/j100301a029
}
