% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{decay_rate_pred}
\alias{decay_rate_pred}
\title{decay_rate_pred}
\usage{
decay_rate_pred(reference_rate, temp_pred, temp_ref, E_a)
}
\arguments{
\item{reference_rate}{known rate for a given temperature}

\item{temp_pred}{temperature for the rate to predict, Kelvin}

\item{temp_ref}{temperature for the reference rate, Kelvin}

\item{E_a}{activation energy of the reaction}
}
\value{
predicted rate at the new temperature
}
\description{
get the predicted inactivation
rate as a function of temperature
based on a known reference rate and
an estimated activation energy.
Assumes constant humidity.
}
