% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{decay_ratio_pred}
\alias{decay_ratio_pred}
\title{decay_ratio_pred}
\usage{
decay_ratio_pred(temp_pred, temp_ref, E_a)
}
\arguments{
\item{temp_pred}{temperature for the rate to predict, Kelvin}

\item{temp_ref}{temperature for the reference rate, Kelvin}

\item{E_a}{activation energy of the reaction}
}
\value{
predicted ratio of the rates (predicted / reference)
}
\description{
get the predicted ratio of inactivation
rates for two temperatures given the
estimated activation energy.
Assumes constant humidity.
}
