% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{density_pure_water}
\alias{density_pure_water}
\title{density_pure_water}
\usage{
density_pure_water(T_kelvin, pressure_atm = 1)
}
\arguments{
\item{T_kelvin}{ambient temperature in Kelvin}

\item{pressure_atm}{ambient pressure in atmospheres, default 1}
}
\description{
Get the predicted density of pure water
at a given temperature and pressure,
using equations from PHREEQC
}
