% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{fraction_change_to_log_concentration_factor}
\alias{fraction_change_to_log_concentration_factor}
\title{fraction_change_to_log_concentration_factor}
\usage{
fraction_change_to_log_concentration_factor(fraction_final, fraction_initial)
}
\arguments{
\item{fraction_final}{final fraction of the component of interest}

\item{fraction_initial}{initial fraction of the component of interest}
}
\value{
natural log of how many times concentrated that quantity has been
}
\description{
convert a change in mass fraction or molar
fraction to a log concentration factor (log fold-concentration
relative to initial)
}
