% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{get_params}
\alias{get_params}
\title{get_params}
\usage{
get_params(parameter_list_name)
}
\arguments{
\item{parameter_list_name}{name of the parameter list to get, one of: \cr

virus_display_names: list of display names for studied viruses \cr

general_palette: general attributes of the project color palette \cr

virus_colors: list of colors associated to each studied virus \cr

material_colors: list of colors associated to each material
on which decay was studied \cr

fluid_type_colors: list of colors associated with human fluids
of interest \cr

treatment_colors: list of colors associated with various
decontamination treatments \cr

general_param_list: general plot styling parameters, such
as line weight and style \cr}
}
\description{
get the desired list of plotting
parameters
}
