% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{half_life_hum}
\alias{half_life_hum}
\title{half_life_hum}
\usage{
half_life_hum(
  reference_half_life,
  temp_pred,
  temp_ref,
  hum_pred,
  hum_ref,
  ERH,
  E_a_dry,
  E_a_wet,
  log_concavity_parameter = 0
)
}
\arguments{
\item{reference_half_life}{known half-life for a
given temperature/humidity}

\item{temp_pred}{temperature for the rate to predict, Kelvin}

\item{temp_ref}{temperature for the reference rate, Kelvin}

\item{hum_pred}{fractional humidity for the rate to predict}

\item{hum_ref}{fractional humidity for the reference rate}

\item{ERH}{fractional effloresecence relative humidity of the
solution containing the virions}

\item{E_a_dry}{activation energy of the reaction below the ERH}

\item{E_a_wet}{activation energy of the reaction above the ERH}

\item{log_concavity_parameter}{log of a parameter governing the
concavity of the relationship between water activity and RH at
RH close to 1. Negative: concave down. Positive: concave up.}
}
\value{
predicted half-life at the new temperature/humidity
}
\description{
get the predicted half-life
as a function of humidity
and temperature based on
a known reference half-life and
activation energy, with the
assumption of an RH^(1/alpha_c)
functional form
}
