% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{half_life_pred}
\alias{half_life_pred}
\title{half_life_pred}
\usage{
half_life_pred(reference_half_life, temp_pred, temp_ref, E_a)
}
\arguments{
\item{reference_half_life}{known half-life for a given temperature}

\item{temp_pred}{temperature for the half-life to predict, Kelvin}

\item{temp_ref}{temperature for the reference half-life, Kelvin}

\item{E_a}{activation energy of the reaction}
}
\value{
predicted half-life at the new temperature
}
\description{
get the predicted half-life
as a function of temperature
based on a known reference half-life and
an estimated activation energy.
Assumes constant humidity.
}
