% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{log_arrenhius_A}
\alias{log_arrenhius_A}
\title{log_arrenhius_A}
\usage{
log_arrenhius_A(k_standard, E_a, T_standard_kelvin)
}
\arguments{
\item{k_standard}{reaction rate at the standard temperature}

\item{E_a}{activation energy for the reaction in J/(Kelvin mol)}

\item{T_kelvin}{value of the standard temperature in Kelvin}
}
\value{
log of the predicted equilibrium salt concencration
}
\description{
Get the log asymptotic reaction rate log(A) rate
given the reaction rate at some standard temperature,
the activation energy E_a, and the value in Kelvin of
the standard temperature
}
