% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{log_arrhenius_rate}
\alias{log_arrhenius_rate}
\title{log_arrhenius_rate}
\usage{
log_arrhenius_rate(log_A, E_a, T_kelvin)
}
\arguments{
\item{log_A}{log of the asymptotic reaction rate at high temperature}

\item{E_a}{activation energy for the reaction in J/(Kelvin mol)}

\item{T_kelvin}{temperature in Kelvin}
}
\value{
log of the predicted equilibrium salt concentration
}
\description{
Get the log reaction rate from the Arrenhius equation
given log asymptotic reaction rate log(A), activation energy
E_a and temperature in kelvin
}
