% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{log_eq_molality_pitzer}
\alias{log_eq_molality_pitzer}
\title{log_eq_molality_pitzer}
\usage{
log_eq_molality_pitzer(
  RH,
  T_kelvin,
  v_salt = 2,
  molality_init_guess = 1,
  molar_mass_solvent = 0.0180153,
  pressure_atm = 1,
  B_0 = B_0_NaCl,
  B_1 = B_1_NaCl,
  C = C_NaCl,
  max_iter = 1000,
  tolerance = 0.01
)
}
\arguments{
\item{RH}{relative humidity}

\item{v_salt}{number of ions, default 2 (NaCl)}

\item{molality_init_guess}{initial guess for the answer, default 1}

\item{molar_mass_solvent}{molar mass (kg/mol) of the solvent, default 0.0180153 (H20)}

\item{pressure_atm}{ambient pressure in atmosphers, default 1}

\item{B_0}{Pitzer B_0 parameter, default 0.07722 (NaCl)}

\item{B_1}{Pitzer B_1 parameter, default 0.25183 (NaCl)}

\item{C}{Pitzer C_parameter, default 0.00106 (NaCl)}

\item{max_iter}{maximum number of iterations to use, default 1000}

\item{tolerance}{tolerance for the log estimate, default 10^(-2)}
}
\value{
solution for the log equilibrium salt concencration
}
\description{
Solve for the log equililbrium salt molality (kg/mol)
as a function of relative humidity by iterating the
\code{\link{log_pred_molality}} function with a practical
osmotic coefficient phi calculated per a simplified
Pitzer equation to find a fixed point
}
