% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{log_pred_molality_pitzer}
\alias{log_pred_molality_pitzer}
\title{log_pred_molality_pitzer}
\usage{
log_pred_molality_pitzer(
  RH,
  phi_salt,
  v_salt = 2,
  molar_mass_solvent = 0.0180153
)
}
\arguments{
\item{RH}{relative humidity}

\item{phi_salt}{practical osmotic coefficient}

\item{v_salt}{number of ions, default 2 (NaCl)}

\item{molar_mass_solvent}{molar mass (kg/mol) of the solvent, default 0.0180153 (H20)}
}
\value{
log of the predicted equilibrium salt concencration
}
\description{
Get approximate log equililbrium salt molality (kg/mol)
as a function of relative humidity and salt constant
product v (num ions) * phi (practical osmotic coefficient)
#'
}
