% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{mass_change_to_log_concentration_factor}
\alias{mass_change_to_log_concentration_factor}
\title{mass_change_to_log_concentration_factor}
\usage{
mass_change_to_log_concentration_factor(
  mass_final,
  mass_initial,
  fraction_initial
)
}
\arguments{
\item{mass_final}{final mass of the solution}

\item{mass_initial}{initial mass of the solution}

\item{fraction}{initial initial mass fraction of the
conserved solution components (solutes, usually)}
}
\value{
natural log of how many times concentrated that quantity has been
}
\description{
convert a change in mass to a log
concentration factor (log fold-concentration
relative to initial)
}
