% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{mass_frac}
\alias{mass_frac}
\title{mass_frac}
\usage{
mass_frac(solute_grams, volume_liters = 1, solvent_density_kg_L = 1)
}
\arguments{
\item{solute_grams}{mass of solute in grams}

\item{volume_liters}{volume of solvent in liters}

\item{solvent_density_kg_L}{density of the solvent in kg per liter}
}
\value{
mass fraction of the solute
}
\description{
Get the mass fraction of a solute (default NaCl)
Makes the assumption that the volume occupied
by the solute can be neglected
}
