% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{mass_frac_to_molality}
\alias{mass_frac_to_molality}
\title{mass_frac_to_molality}
\usage{
mass_frac_to_molality(mass_frac, solute_molar_mass_grams = 58.44)
}
\arguments{
\item{solute_molar_mass_grams}{molar mass of the solute in
grams (default 58.44 (NaCl))}

\item{mass_fract}{mass fraction of solute}
}
\value{
molality of the solute
}
\description{
Get the molality of a solute (default NaCl in water)
given the mass fraction
}
