% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{mass_frac_to_molar_frac}
\alias{mass_frac_to_molar_frac}
\title{mass_frac_to_molar_frac}
\usage{
mass_frac_to_molar_frac(
  mass_frac,
  molar_mass = 0.05844,
  molar_mass_others = 0.0180153
)
}
\arguments{
\item{mass_frac}{mass fraction of the component of interests}

\item{molar_mass}{molar mass of the component of interest,
default 0.05844 mol/kg (NaCl)}

\item{molar_mass_others}{average molar mass of other components,
default 0.0180153 mol/kg (pure H20)}
}
\description{
convert a mass fraction to a corresponding
molar fraction (default NaCl in water)
}
