% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{mechanistic_decay_rate}
\alias{mechanistic_decay_rate}
\title{mechanistic_decay_rate}
\usage{
mechanistic_decay_rate(
  relative_humidity,
  T_kelvin,
  log_concentration_term,
  ERH,
  log_A_dry,
  E_a_dry,
  log_A_wet,
  E_a_wet
)
}
\arguments{
\item{relative_humidity}{ambient relative humidity as a decimal}

\item{T_kelvin}{ambient temperature in Kelvin}

\item{log_concentration_term}{log of the concentration term modifying the reaction rate}

\item{ERH}{efflorescence relative humidity for the inactivating agent(s)}

\item{log_A_dry}{log of the asymptotic reaction rate at high temperatures but below the ERH (i.e. when the agent(s) that are leading order virus inactivators in solution effloresce)}

\item{E_a_dry}{activation energy of the reaction below the ERH (i.e. when the agent(s) that are leading order virus inactivators in solution effloresce)}

\item{log_A_wet}{log of the asymptotic reaction rate at high temperatures above the ERH (i.e. in solution) with asymptotically high concentration of inactivating agents(s)}

\item{E_a_wet}{activation energy of the reaction above the ERH (i.e. in solution)}
}
\value{
predicted exponetial decay rate of virus (reaction rate k of the inactivation reaction)
}
\description{
Get the predicted decay rate (reaction rate of
the inactivation reaction) of viable virus
given temperature and humidity and the
ERH and mass fraction of the inactivating
reactant (salt or something else).
By default, mass fractions are converted
into molarities according to the function
\link{mass_frac_to_molarity}
}
