% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{molality}
\alias{molality}
\title{molality}
\usage{
molality(
  solute_grams,
  volume_liters = 1,
  solvent_density_kg_L = 1,
  solute_molar_mass_grams = 58.44
)
}
\arguments{
\item{solute_grams}{mass of solute in grams}

\item{volume_liters}{volume of solvent in liters}

\item{solvent_density_kg_L}{density of the solvent in kg per liter}

\item{solute_molar_mass_grams}{molar mass of the solute in g/mol}
}
\value{
molality of the solution
}
\description{
Get the molality of a solute (default NaCl).
Makes the assumption that the volume occupied
by the solute can be neglected
}
