% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{molality_to_mass_frac}
\alias{molality_to_mass_frac}
\title{molality_to_mass_frac}
\usage{
molality_to_mass_frac(molality, solute_molar_mass_grams = 58.44)
}
\arguments{
\item{molality}{molality of solute}

\item{solute_molar_mass_grams}{molar mass of the solute in
grams (default 58.44 (NaCl))}
}
\value{
mass fraction of the solute
}
\description{
Get the mass fraction of a solute (default NaCl in water)
given the molality
}
