% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{pitzer_phi}
\alias{pitzer_phi}
\title{pitzer_phi}
\usage{
pitzer_phi(molality, T_kelvin, B_0, B_1, C, pressure_atm = 1)
}
\arguments{
\item{molality}{molality of solute (moles / kg solvent)}

\item{T_kelvin}{ambient temperature in Kelvin}

\item{B_0}{Pitzer B_0 parameter}

\item{B_1}{Pitzer B_1 parameter}

\item{C}{Pitzer C_parameter}

\item{pressure_atm}{ambient pressure in atmospheres, default 1}
}
\value{
practical osmotic coefficient Phi
}
\description{
Solve the Pitzer equation for the
practical osmotic coefficient Phi at
a given temperature and molality for
a 1-1 solute
}
