% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{pitzer_phi_nacl}
\alias{pitzer_phi_nacl}
\title{pitzer_phi_nacl}
\usage{
pitzer_phi_nacl(molality, T_kelvin, pressure_atm = 1)
}
\arguments{
\item{molality}{molality of NaCl (moles / kg solvent)}

\item{T_kelvin}{ambient temperature in Kelvin}

\item{pressure_atm}{ambient pressure in atmospheres, default 1}
}
\value{
practical osmotic coefficient Phi
}
\description{
Solve the Pitzer equation for the
practical osmotic coefficient Phi of
NaCl at a given temperature and molality
}
