% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{predict_log_concentration_factor}
\alias{predict_log_concentration_factor}
\title{predict_log_concentration_factor}
\usage{
predict_log_concentration_factor(
  relative_humidity,
  log_shape_parameter_concavity,
  log_shape_parameter_steepness,
  initial_mass_fraction_solute
)
}
\arguments{
\item{relative_humidity}{ambient relative humidity}

\item{log_shape_parameter_concavity}{parameter governing
concavity of the mole-fraction / water activity relationship (negative
= concave down, positive = concave up)
log_shape_parameter_steepness parameter governing
steepness of the mole-fraction / water activity relationship (larger values = steeper)}

\item{initial_mass_fraction_solute}{mass fraction of the
solute(s) in the solution at t = 0.}
}
\value{
log concentration factor for the solution
}
\description{
predict (the natural log of)
how many-fold concentrated a solution
will be at equilibrium given
the initial mass fraction,
two shape parameters governing
non-ideality
and the initial solute concentration
}
