% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{predict_titers_explicit_evaporation}
\alias{predict_titers_explicit_evaporation}
\title{predict_titers_explicit_evaporation}
\usage{
predict_titers_explicit_evaporation(
  time,
  intercept,
  initial_decay_rate,
  eq_decay_rate,
  normalized_evap_rate_B,
  equilibrium_concentration_factor
)
}
\arguments{
\item{time}{time since depsotion at which the titer is observed}

\item{intercept}{log10 titer at time = 0}

\item{initial_decay_rate}{decay rate of viable virus at t = 0}

\item{eq_decay_rate}{exponential decay rate of viable virus after evaporative equilibrium is reached}

\item{normalized_evap_rate_B}{normalized linear rate B of evaporation, in the same units as time (so that full evaporation would occur at t = 1 / B}

\item{equilibrium_concentration_factor}{degree of concentration of the solution (relative to the initial) at equilibrium (sets the point at which evaporation stops)}
}
\value{
predicted log10 titer
}
\description{
predict log10 titer as a function of time with
explicit modeling of evaporation during
the transient phase
}
