% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{predict_titers_exponential_decay}
\alias{predict_titers_exponential_decay}
\title{predict_titers_exponential_decay}
\usage{
predict_titers_exponential_decay(time, intercept, decay_rate)
}
\arguments{
\item{time}{time since depsotion at which the titer is observed}

\item{intercept}{log10 titer at time = 0}

\item{decay_rate}{exponential decay rate of viable virus}
}
\value{
predicted log10 titer
}
\description{
predict log10 titer as a function of time under
the assumption of simple first-order kinetics
(exponential decay)
}
