% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{predict_titers_implicit_evaporation}
\alias{predict_titers_implicit_evaporation}
\title{predict_titers_implicit_evaporation}
\usage{
predict_titers_implicit_evaporation(
  time,
  intercept,
  transient_decay_rate,
  eq_decay_rate,
  drying_time
)
}
\arguments{
\item{time}{time since depsotion at which the titer is observed}

\item{intercept}{log10 titer at time = 0}

\item{transient_decay_rate}{exponential decay rate of viable virus
prior to evaporative equilibrium}

\item{eq_decay_rate}{exponential decay rate of viable virus at
evaporative equilibrium}

\item{drying_time}{time evaporative equilibrium is reached}
}
\value{
predicted log10 titer
}
\description{
predict log10 titer as a function of time under
the assumption of two phases of
simple first-order kinetics
(exponential decay): a transient phase prior
to evaporative equilibrium and an equilibrium
phase thereafter
}
