% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{read_data_for_plotting}
\alias{read_data_for_plotting}
\title{read_data_for_plotting}
\usage{
read_data_for_plotting(data_path)
}
\arguments{
\item{data_path}{path to data to read in}
}
\value{
tibble of formatted data
}
\description{
function to read in cleaned data and format
it as a tibble that can be used for plotting
}
