% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{save_plot}
\alias{save_plot}
\title{save_plot}
\usage{
save_plot(outpath, fig, ...)
}
\arguments{
\item{outpath}{path to save the figure to}

\item{fig}{ggplot figure to save}

\item{...}{other parameters passed to cowplot::save_plot()}
}
\value{
nothing
}
\description{
override cowplot's save_plot
to include font embedding
}
