% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_color_virus}
\alias{scale_color_virus}
\title{scale_color_virus}
\usage{
scale_color_virus(...)
}
\arguments{
\item{...}{parameters passed to scale_color_manual()}
}
\value{
color scale
}
\description{
default color scale for viruses
for this project
}
