% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_virus}
\alias{scale_fill_virus}
\title{scale_fill_virus}
\usage{
scale_fill_virus(...)
}
\arguments{
\item{...}{parameters passed to scale_fill_manual()}
}
\value{
fill scale
}
\description{
default fill scale for viruses
for this project
}
