% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_shape_virus}
\alias{scale_shape_virus}
\title{scale_shape_virus}
\usage{
scale_shape_virus(...)
}
\arguments{
\item{...}{parameters passed to scale_shape_manual()}
}
\value{
shape scale
}
\description{
default shape scale for viruses
for this project
}
