% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_size_virus}
\alias{scale_size_virus}
\title{scale_size_virus}
\usage{
scale_size_virus(base_size = 1, shrinkage = 0.75, aesthetics = "size", ...)
}
\arguments{
\item{base_size}{size for the non-triangle points}

\item{shrinkage}{how much to shrink the triangles
relative to the other shapes (default 0.75)}

\item{aesthetics}{aesthetics for which the scale
will be used (default size)}

\item{...}{other parameters passed to scale_discrete_manual()}
}
\value{
size scale
}
\description{
default size scale for viruses
for this project to accompany
scale_shape_virus, since the
triangles (HCoV-229E and HCoV-OC43)
are larger than the other shapes
unless resized
}
