% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_x_log10_mathformat}
\alias{scale_x_log10_mathformat}
\title{scale_x_log10_mathformat}
\usage{
scale_x_log10_mathformat(...)
}
\arguments{
\item{...}{arguments passed to scale_x_continuous()}
}
\value{
a scale function
}
\description{
convenience scale function
for putting x breaks on
powers of 10 in ggplot
}
