% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_y_log10_mathformat}
\alias{scale_y_log10_mathformat}
\title{convenience scale function
for putting y breaks on
powers of 10 in ggplot}
\usage{
scale_y_log10_mathformat(...)
}
\arguments{
\item{...}{arguments passed to scale_y_continuous()}
}
\value{
a scale function
}
\description{
convenience scale function
for putting y breaks on
powers of 10 in ggplot
}
