% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{solute_to_solvent_mole_ratio}
\alias{solute_to_solvent_mole_ratio}
\title{solute_to_solvent_mole_ratio}
\usage{
solute_to_solvent_mole_ratio(
  relative_humidity,
  log_shape_parameter_concavity,
  log_shape_parameter_steepness
)
}
\arguments{
\item{relative_humidity}{ambient relative humidity}

\item{log_shape_parameter_concavity}{parameter governing
concavity of the mole-fraction / water activity relationship (negative
= concave down, positive = concave up)
log_shape_parameter_steepness parameter governing
steepness of the mole-fraction / water activity relationship near a_w/rh = 1(larger values = steeper)}
}
\value{
predicted ratio of solute mole fraction to
solvent mole fraction at equilibrium
}
\description{
predict the ratio of solute mole
fraction to solvent mole fraction
at evaporative equilibrium for a
given relative humidity, using a
two-parameter semi-mechanistic
functional form
}
