% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{solve_quartic}
\alias{solve_quartic}
\title{solve_quartic}
\usage{
solve_quartic(a_vec)
}
\arguments{
\item{a_vec}{vector of coefficients in order a0, a1, a2, a3, a4 \cr
where the quartic is a4 x^4 + a3 x^3 + a2 x^2 + a1 x + a0}
}
\value{
the four roots
}
\description{
analytically sole a quartic equation with the given coefficients
}
