% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\docType{data}
\name{strain_order}
\alias{strain_order}
\title{strain_order}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
strain_order
}
\description{
order in which to label virus strains
}
\keyword{datasets}
