% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{tang_aw}
\alias{tang_aw}
\title{tang_aw}
\usage{
tang_aw(mass_frac)
}
\arguments{
\item{mass_fract}{mass fraction of NaCl}
}
\value{
corresponding water activity
}
\description{
Get the water activity of
NaCl in pure water at env conditions
corresponding to the given mass fraction
using the empirical equation of Tang 1996
https://doi.org/10.1029/96JD03003
}
