% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{tang_mass_frac}
\alias{tang_mass_frac}
\title{tang_mass_frac}
\usage{
tang_mass_frac(RH)
}
\arguments{
\item{RH}{ambient relative humidity, as a decimal}
}
\value{
mass fraction of NaCl
}
\description{
Get the mass fraction of NaCl
NaCl in pure water at the given RH
by inverting the empirical equation of Tang 1996
https://doi.org/10.1029/96JD03003
analytically using the quartic formula
}
