% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{theme_project}
\alias{theme_project}
\title{theme_project}
\usage{
theme_project(base_size = 30, ...)
}
\arguments{
\item{base_size}{base font size for the theme, passed to
theme_classic (default 30)}

\item{...}{other parameters passed to theme_classic()}
}
\value{
theme
}
\description{
variant of theme_classic() ggplot theme for
this project
}
