% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_kinetics.R
\name{to_AH}
\alias{to_AH}
\title{to_AH}
\usage{
to_AH(RH, T_kelvin)
}
\arguments{
\item{RH}{relative humidity}

\item{T_kelvin}{Ambient temperature in Kelvin}
}
\value{
corresponding absolute humidity in g/m^3
}
\description{
convert a value of RH to its AH given
the ambient temperature T_kelvin
}
