% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_constants.R
\name{to_Celsius}
\alias{to_Celsius}
\title{to_Celsius}
\usage{
to_Celsius(temp_K)
}
\arguments{
\item{temp_K}{temperature in Kelvin}
}
\value{
temperature in Celsius
}
\description{
convert a temperature in Kelvin to Celsius
}
