% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{to_decay_rate}
\alias{to_decay_rate}
\title{to_decay_rate}
\usage{
to_decay_rate(fractional_life, fraction = 1/2, decay_rate_log_base = 10)
}
\arguments{
\item{fractional_life}{fractional life to convert
in a given time unit}

\item{fraction}{used (default 1 / 2: half-life)}

\item{decay_rate_log_base}{base of the logarithm
used to calculate the decay rate (default 10)}
}
\value{
decay_ rate in logs of the given
base per time unit
}
\description{
convert a an nth-life
(i.e. how long until the fraction in
question is all that remains)
into a decay rate in logs per hour
(default log base 10)
}
