% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{to_fractional_life}
\alias{to_fractional_life}
\title{to_fractional_life}
\usage{
to_fractional_life(decay_rate, fraction, decay_rate_log_base = 10)
}
\arguments{
\item{decay_rate}{exponential decay rate}

\item{fraction}{fraction to calculate for}

\item{decay_rate_log_base}{base of the logarithm
used to calculate the decay rate (default 10)}
}
\value{
fractional life for the given fraction
}
\description{
convert a decay rate to an nth-life
(i.e. how long until the fraction in
question is all that remains)
}
