% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_kinetics.R
\name{to_half_life}
\alias{to_half_life}
\title{to_half_life}
\usage{
to_half_life(decay_rate, decay_rate_log_base = 10)
}
\arguments{
\item{decay_rate}{exponential decay rate}

\item{decay_rate_log_base}{base of the logarithm
used to calculate the decay rate (default 10)}
}
\value{
fractional life for the given fraction
}
\description{
convert a decay rate to a half_life
}
