% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_constants.R
\name{to_Kelvin}
\alias{to_Kelvin}
\title{to_Kelvin}
\usage{
to_Kelvin(temp_C)
}
\arguments{
\item{temp_C}{temperature in Celsius}
}
\value{
temperature in Kelvin
}
\description{
convert a temperature in Celsius to Kelvin
}
