% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\docType{data}
\name{virus_order}
\alias{virus_order}
\title{virus_order}
\format{
An object of class \code{character} of length 6.
}
\usage{
virus_order
}
\description{
order in which to label viruses
}
\keyword{datasets}
